+++ api/src/compiler.ts
@@ -1,6 +1,3 @@
-import { Target } from "./types";
-import { ensure } from "./utils";
-
 const solc = require('solc');
 
 export interface Contracts {
@@ -21,7 +18,6 @@ interface SolcContracts {
       enabled: true;
       runs: number;
     };
-    evmVersion?: Target;
     outputSelection: {
       "*": {
         '*': string[];
@@ -37,13 +33,10 @@ const toCompilerContracts = (contracts: Contracts): CompilerContracts =>
     {} as CompilerContracts
   );
 
-const evmVersion = (version: Target) => version === "london" ? {} : {evmVersion: version};
-
-const prepareSolcContracts = (contracts: Contracts, target: Target): SolcContracts => ({
+const prepareSolcContracts = (contracts: Contracts): SolcContracts => ({
   language: 'Solidity',
   sources: toCompilerContracts(contracts),
   settings: {
-    ...evmVersion(target),
     outputSelection: {
       '*': {
         '*': ['*']
@@ -52,14 +45,13 @@ const prepareSolcContracts = (contracts: Contracts, target: Target): SolcContrac
   }
 });
 
-const prepareOptimizedSolcContracts = (contracts: Contracts, runs: number, target: Target): SolcContracts => ({
+const prepareOptimizedSolcContracts = (contracts: Contracts, runs: number): SolcContracts => ({
   language: 'Solidity',
   sources: toCompilerContracts(contracts),
   settings: {
-    ...evmVersion(target),
     optimizer: {
-      runs,
       enabled: true,
+      runs: runs
     },
     outputSelection: {
       '*': {
@@ -70,24 +62,8 @@ const prepareOptimizedSolcContracts = (contracts: Contracts, runs: number, targe
 });
 
 const preprocessBytecode = (bytecode: string): string => {
-  let filteredBytecode = "";
-  const start = bytecode.indexOf('6080604052');
-  //
-  // metadata separator (solc >= v0.6.0)
-  //
-  const ipfsMetadataEnd = bytecode.indexOf('a264697066735822');
-  filteredBytecode = bytecode.slice(start, ipfsMetadataEnd);
-
-  //
-  // metadata separator for 0.5.16
-  //
-  const bzzr1MetadataEnd = filteredBytecode.indexOf('a265627a7a72315820');
-  filteredBytecode = filteredBytecode.slice(0, bzzr1MetadataEnd);
-
-  // debug
-  // logger.info(loggerOptions, `processed bytecode: ${bytecode}`);
-
-  return filteredBytecode;
+  // TODO preprocess bytecode
+  return bytecode;
 }
   
 const loadCompiler = async (version: string): Promise<any> => (
@@ -99,50 +75,21 @@ const loadCompiler = async (version: string): Promise<any> => (
 )
 
 interface CompilerResult {
-  errors?: {
-    type: string;
-    formattedMessage: string;
-  }[];
-  contracts?: {
-    [filename: string]: {
-      [name: string]: any;
-    };
-  };
-  sources: {
-    [filename: string]: {
-      id: number;
-    };
-  };
-}
-
-const compressErrors = ({errors=[]}: CompilerResult): string => errors
-  .reduce((prev, error) => (error.type.toLowerCase().includes("error")
-    ? `${prev}\n${error.formattedMessage}`
-    : prev
-  ), "");
-
-// type Bytecode = string;
-interface Compile {
   abi: any;
   bytecode: string;
 }
-export const compileContracts = async (contractName: string, contractFilename: string, source: string, version: string, target: Target, optimizer?: boolean, runs=200): Promise<Compile> => {
+export const compileContracts = async (contractName: string, contractFilename: string, source: string, version: string, optimizer?: boolean, runs=200): Promise<CompilerResult> => {
   const compiler = await loadCompiler(version);
   const contracts = JSON.parse(source);
-  
   const solcData = optimizer 
-    ? prepareOptimizedSolcContracts(contracts, runs, target)
-    : prepareSolcContracts(contracts, target);
+    ? prepareOptimizedSolcContracts(contracts, runs)
+    : prepareSolcContracts(contracts);
 
-  const compilerResult = JSON.parse(compiler.compile(JSON.stringify(solcData)));
-  const error = compressErrors(compilerResult);
-  ensure(error === "", error);
-  ensure(contractFilename in compilerResult.contracts, "Filename does not exist in compiled results");
-  ensure(contractName in compilerResult.contracts[contractFilename], "Name does not exist in compiled results");
-  const result = compilerResult.contracts[contractFilename][contractName];
-  // return preprocessBytecode(result.evm.bytecode.object);
+  const compilerResult = JSON.parse(compiler.compile(JSON.stringify(solcData)))
+  const result = compilerResult.contracts[contractFilename][contractName].evm;
+  console.log(result);
   return {
-    abi: result.abi,
-    bytecode: preprocessBytecode(result.evm.bytecode.object)
+    abi: result.abi.object,
+    bytecode: preprocessBytecode(result.bytecode.object)
   }
 }
