+++ api/src/connector.ts
@@ -6,20 +6,27 @@ export const config = {
   httpPort: process.env.PORT || 8000,
   nodeWs: 'ws://substrate-node:9944',
   postgresConfig: {
-    // port: 54321,
-    // host: '0.0.0.0',
+    host: '0.0.0.0',
     user: process.env.POSTGRES_USER || 'reefexplorer',
-    host: process.env.POSTGRES_HOST || 'postgres',
+    // host: process.env.POSTGRES_HOST || 'postgres',
     database: process.env.POSTGRES_DATABASE || 'reefexplorer',
     password: process.env.POSTGRES_PASSWORD || 'reefexplorer',
-    port: parseInt(process.env.POSTGRES_PORT, 10) || 5432,
+    port: 54321,
+    // port: parseInt(process.env.POSTGRES_PORT, 10) || 5432,
   }
+}
+
+export const connect = async (): Promise<Pool> => {
+  const connection = new Pool({...config.postgresConfig});
+  await connection.connect()
+  return connection;
 }
 
 export const query = async <Res>(statement: string, args: any[]): Promise<Res[]> => {
   const db = new Pool({...config.postgresConfig});
   await db.connect();
   const result = await db.query(statement, [...args]);
+  await db.end();
   return result.rows;
 }
 
@@ -44,10 +51,18 @@ export const getReefPrice = async (): Promise<Price> => axios
   });
   
   
-export const authenticationToken = async (token: string): Promise<boolean> => await axios
-  .get(`https://www.google.com/recaptcha/api/siteverify?secret=${config.recaptchaSecret}&response=${token}`)
-  .then((res) => res.data.success)
-  .catch((err) => {
-    console.log(err);
-    throw new Error("Can not extract recaptcha token...")
-  });
+  const googleAuthenticatorApi = axios.create({
+    baseURL: "`https://www.google.com/recaptcha/",
+    headers: {
+      "Content-Type": "application/x-www-form-urlencoded; charset=utf-8"
+    },
+  })
+  export const authenticationToken = async (token: string): Promise<boolean> => await googleAuthenticatorApi
+    .post(`api/siteverify?secret=${config.recaptchaSecret}&response=${token}`, {})
+    .then((res) => res.data)
+    .then((res: any) => res.text()) // TODO any is a hack!
+    .then((res) => JSON.parse(res).success)
+    .catch((err) => {
+      console.log(err);
+      throw new Error("Can not extract recaptcha token...")
+    });
