+++ api/src/utils.ts
@@ -1,29 +1,13 @@
 
-export class StatusError extends Error {
-  status: number;
-
-  constructor(message: string, status: number) {
-    super(message);
-    this.status = status;
-  }
-}
-
-export const ensure = (condition: boolean, message: string, status=400): void => {
+export const ensure = (condition: boolean, message: string): void => {
   if (!condition) {
-    throw new StatusError(message, status);
+    throw new Error(message);
   }
 }
 
 export const ensureObjectKeys = <Object extends {}, K extends keyof Object> (obj: Object, keys: K[]): void => {
   keys
     .forEach((key) => 
-      ensure(!!obj[key], `Parameter ${key} is missing`, 400)
+      ensure(!!obj[key], `Parameter ${key} is missing`)
     );
-}
-
-export const errorStatus = (err: any): number => {
-  if (err instanceof StatusError) {
-    return err.status;
-  }
-  return 400;
 }
